refactor: extract common elements to resolve circular dependencies
