+++ backend/app/schemas/match_events.py
@@ -1,9 +1,9 @@
 from pydantic import BaseModel, ConfigDict
 
 from app.models.match_event import EventType
+from app.schemas.common import TeamSummary
 from app.schemas.metadata import ResponseMetadata
 from app.schemas.players import PlayerSummary
-from app.schemas.teams import TeamSummary
 
 
 class MatchEventItemResponse(BaseModel):
