+++ backend/app/schemas/teams.py
@@ -1,15 +1,6 @@
 from pydantic import BaseModel, ConfigDict
 
-from .standings import TeamStandingsSummary
-
-
-class TeamSummary(BaseModel):
-    id: int
-    name: str
-    short_name: str
-    logo_url: str | None = None
-
-    model_config = ConfigDict(from_attributes=True)
+from app.schemas.common import TeamStandingsSummary, TeamSummary
 
 
 class TeamProfileResponse(BaseModel):
